ref: Streamline refreshPackages
